backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.3.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
tensorflow==2.13.0  # or pytorch if using PyTorch
scikit-learn==1.3.2
python-dotenv==1.0.0
